# [chore] [exporter/splunkhec] Split the profiling data before processing