--- a/exporter/splunkhecexporter/batchperscope_test.go
+++ b/exporter/splunkhecexporter/batchperscope_test.go
@@ -21,18 +21,28 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/consumer/consumertest"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+	"go.uber.org/zap/zaptest/observer"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/golden"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/pdatatest/plogtest"
 )
 
 func TestBatchLogs_ConsumeLogs(t *testing.T) {
+	type debugMsg struct {
+		text         string
+		droppedCount int64
+	}
+	profilingDropped := debugMsg{text: "Profiling data is not allowed", droppedCount: 4}
+	logsDropped := debugMsg{text: "Log data is not allowed", droppedCount: 5}
 	tests := []struct {
 		name             string
 		profilingEnabled bool
 		logsEnabled      bool
 		in               string
 		out              []string
+		wantDropped      []debugMsg
 	}{
 		{
 			name:             "profiling_only_both_enabled",
@@ -52,6 +62,7 @@ func TestBatchLogs_ConsumeLogs(t *testing.T) {
 			logsEnabled: true,
 			in:          "profiling_only.yaml",
 			out:         []string{},
+			wantDropped: []debugMsg{profilingDropped},
 		},
 		{
 			name:             "regular_logs_only_both_enabled",
@@ -71,6 +82,7 @@ func TestBatchLogs_ConsumeLogs(t *testing.T) {
 			profilingEnabled: true,
 			in:               "regular_logs_only.yaml",
 			out:              []string{},
+			wantDropped:      []debugMsg{logsDropped},
 		},
 		{
 			name:             "combined_both_enabled",
@@ -84,26 +96,27 @@ func TestBatchLogs_ConsumeLogs(t *testing.T) {
 			logsEnabled: true,
 			in:          "combined.yaml",
 			out:         []string{"regular_logs_only.yaml"},
+			wantDropped: []debugMsg{profilingDropped},
 		},
 		{
 			name:             "combined_logs_disabled",
 			profilingEnabled: true,
 			in:               "combined.yaml",
 			out:              []string{"profiling_only.yaml"},
-		},
-		{
-			name: "combined_both_disabled",
-			in:   "combined.yaml",
-			out:  []string{},
+			wantDropped:      []debugMsg{logsDropped},
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			sink := &consumertest.LogsSink{}
+			core, obs := observer.New(zapcore.DebugLevel)
+			logger := zap.New(core)
+
 			consumer := &perScopeBatcher{
 				profilingEnabled: tt.profilingEnabled,
 				logsEnabled:      tt.logsEnabled,
+				logger:           logger,
 				next:             sink,
 			}
 
@@ -113,13 +126,19 @@ func TestBatchLogs_ConsumeLogs(t *testing.T) {
 			err = consumer.ConsumeLogs(context.Background(), logs)
 			assert.NoError(t, err)
 
-			assert.Equal(t, len(tt.out), len(sink.AllLogs()))
+			require.Equal(t, len(tt.out), len(sink.AllLogs()))
 			for i, out := range tt.out {
 				expected, err := golden.ReadLogs("testdata/batchperscope/" + out)
 				require.NoError(t, err)
 				assert.NoError(t, plogtest.CompareLogs(expected, sink.AllLogs()[i]))
 			}
+
+			require.Equal(t, len(tt.wantDropped), obs.Len())
+			for _, entry := range tt.wantDropped {
+				filtered := obs.FilterMessage(entry.text)
+				require.Equal(t, 1, filtered.Len())
+				assert.Equal(t, entry.droppedCount, filtered.All()[0].ContextMap()["dropped_records"])
+			}
 		})
 	}
-
 }